+++ Editor/Scripts/Tools/PaintTool.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using InstancePainter.Runtime;
+using InstancePainter;
 using UnityEditor;
 using UnityEngine;