+++ Editor/Scripts/Tools/RectTool.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using InstancePainter.Runtime;
+using InstancePainter;
 using UnityEditor;
 using UnityEngine;
 using Random = UnityEngine.Random;