+++ Editor/Scripts/Windows/IPEditorWindow.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using InstancePainter.Runtime;
+using InstancePainter;
 using InstancePainter.Editor;
 using UnityEditor;
 using UnityEngine;